Create Nikhilam_and_Nines
Learning Veda Math can be daunting if you're not familiar with the mental steps needed to come to the correct number. With nines as the multiplier, there are three types. This program will walk you through the steps, tell you which type of problem it is and perform the intermediary steps of finding other numbers from the multiplicand. I've only performed limited tests on this.